Make the base directory for requests configurable
The `/requests` path was hardcoded in a bunch of places making
it difficult to install fmperf in a running container where
you don't have root access to `mkdir /create`.

I've added an environment variable that defaults to ""
so that all paths become local. If the path is not empty
the code makes sure that there is an "/" at the end to
concatenate the paths.

To prevent breaking existing scripts, in the Dockerfile
the environment variable is set to `/requests`

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>